Pojects_Exercises/potato_diseases_project/gcp/requirements.txt
tensorflow==2.18.0
google-cloud-storge==1.16.1
pillow==6.00